updated server wrapper for album details
originally i used the already existing Album object, which holds an Artist object (required for another endpoint of the API), but in this case the server only returns a string.
the solution is to mimic the server wrapper as it is, and convert it to the "normal" Album object somewhere else in the code.
this unfortunately means some code duplication