removed id field from some classes
the id field is unreliable. it is sometimes empty, or even missing. so instead, i will use names as id's (the endpoints use names anyways).
names can also be empty or missing, so worst case we are at the same place as we were with the id field.
given the above, there is no point in keeping the id field