limit the search results to 100
instead of implementing paging, we limit the search query to 100 results.
not only will this mean less code (and therefore potential bugs), but it also makes sense because if the user cannot find the artist they were looking for in the first 100 results, they probably need to adjust their search query.